Time: 470 ms (5.04%), Space: 318.9 MB (5.04%) - LeetHub

+++ 650-2-keys-keyboard/650-2-keys-keyboard.java
@@ -16,7 +16,7 @@ class Solution {
         
     }
     
-    public long memo(int screen ,int buffer , int dest , long[][] dp ){
+    public long memo(int screen ,int buffer , long dest , long[][] dp ){
         
         if( screen > dest  ) return Integer.MAX_VALUE;
         if( screen == dest ) return 0;
